Faster updates
Instead of looping the document n times (n being the number of gremlins), we create a regexp that matches every gremlin and only loop the document once.
With 4 gremlins, that makes it about 3 times faster [1]. As we add more gremlins, it should not go slower with this algorithm.

I've also done some other cosmetic changes:
- Change some variable names
- Extract functions

[1] ~40ms instead of ~160ms on my laptop for the HTML of https://html.spec.whatwg.org